requirements.txt
fastapi==0.61.1
pydantic==1.6.1
uvicorn==0.11.8
motor==2.2.0
pydantic[email]
python-decouple
email-validator
PyJWT==1.7.1
passlib
bcrypt
functools
